refactor: 장소 상세 페이지 Section 컴포넌트를 별도 파일로 분리
