+++ src/views/detail-place/index.tsx
@@ -16,6 +16,7 @@ import allReview from '@/src/assets/images/all_review_icon.svg'
 import kakaoReview from '@/src/assets/images/kakao_review_icon.svg'
 import { useRouter } from 'next/navigation'
 import { ReviewStats } from '@/src/features'
+import { Section } from './section'
 
 export default function DetailPlace({ id }: { id: string }) {
   const { data: placeData } = useGetPlace(id)
@@ -152,118 +153,91 @@ export default function DetailPlace({ id }: { id: string }) {
           <Spacer height={20} />
         </div>
 
-        <Section
-          title='리뷰'
-          subtitle='가장 언급 많은 키워드 랭킹이에요!'
-          button={
-            <span
-              className='text-sub text-gray-400 cursor-pointer'
-              onClick={() => router.push(`/places/${id}/reviews/new`)}
-            >
-              작성하기
-            </span>
-          }
-          ref={reviewRef}
-        >
-          <Spacer height={15} />
-          {reviewData.length === 0 ? (
-            <ReviewStats
-              placeOnLineReviewStats={placeData.place_one_line_review_stats}
-              AverageRating={placeData.average_rating}
-            />
-          ) : (
-            <div className='h-[100px] flex items-center justify-center'>
-              <span className='text-description text-sub'>
-                아직 리뷰를 기다리고 있어요!
-              </span>
-            </div>
-          )}
-          <Spacer height={15} />
-        </Section>
-
-        {reviewData.length === 0 && (
-          <>
-            <Spacer height={20} />
-            <Spacer height={8} className='bg-light-gray' />
-
-            <div className='flex flex-col w-full gap-[25px] px-[20px]'>
-              <Spacer height={4} />
-              {reviewData.map((review) => (
-                <ReviewCommentCard
-                  key={review.id}
-                  id={placeData.id.toString()}
-                  content={review}
-                  refetch={refetch}
-                />
-              ))}
-              <Spacer height={24} />
-            </div>
-
-            <Spacer height={8} className='bg-light-gray' />
-            <Spacer height={20} />
-
-            <div className='flex flex-col justify-center items-center gap-[18px]'>
-              <Link href={`/places/${id}/reviews`}>
-                <div className='w-[315px] relative h-[45px] flex flex-row items-center justify-start py-0 pl-[64px] gap-[14.73px] shadow-[0px_0px_5.1px_rgba(0,_0,_0,_0.4)] rounded-[99px] overflow-hidden'>
-                  <div className='font-bold w-[120px] text-main flex items-center'>
-                    전체 리뷰 바로보기
-                  </div>
-                  <div className='w-[75px] h-[75px] relative flex justify-center items-center'>
-                    <Image className='relative' fill alt='' src={allReview} />
-                  </div>
-                </div>
-              </Link>
-
-              <Link
-                href={`https://place.map.kakao.com/m/${placeData.kakao_place_id}`}
+        <div ref={reviewRef} className='w-full flex flex-col items-center'>
+          <Section
+            title='리뷰'
+            subtitle='가장 언급 많은 키워드 랭킹이에요!'
+            button={
+              <span
+                className='text-sub text-gray-400 cursor-pointer'
+                onClick={() => router.push(`/places/${id}/reviews/new`)}
               >
-                <div className='w-[315px] relative h-[45px] flex flex-row items-center justify-start py-0 pl-[64px] gap-[14.73px] shadow-[0px_0px_5.1px_rgba(0,_0,_0,_0.4)] rounded-[99px] overflow-hidden'>
-                  <div className='font-bold w-[132px] text-main flex items-center'>
-                    카카오 리뷰 바로가기
+                작성하기
+              </span>
+            }
+          >
+            <Spacer height={15} />
+            {reviewData.length === 0 ? (
+              <ReviewStats
+                placeOnLineReviewStats={placeData.place_one_line_review_stats}
+                AverageRating={placeData.average_rating}
+              />
+            ) : (
+              <div className='h-[100px] flex items-center justify-center'>
+                <span className='text-description text-sub'>
+                  아직 리뷰를 기다리고 있어요!
+                </span>
+              </div>
+            )}
+            <Spacer height={15} />
+          </Section>
+
+          {reviewData.length === 0 && (
+            <>
+              <Spacer height={20} />
+              <Spacer height={8} className='bg-light-gray' />
+
+              <div className='flex flex-col w-full gap-[25px] px-[20px]'>
+                <Spacer height={4} />
+                {reviewData.map((review) => (
+                  <ReviewCommentCard
+                    key={review.id}
+                    id={placeData.id.toString()}
+                    content={review}
+                    refetch={refetch}
+                  />
+                ))}
+                <Spacer height={24} />
+              </div>
+
+              <Spacer height={8} className='bg-light-gray' />
+              <Spacer height={20} />
+
+              <div className='flex flex-col justify-center items-center gap-[18px]'>
+                <Link href={`/places/${id}/reviews`}>
+                  <div className='w-[315px] relative h-[45px] flex flex-row items-center justify-start py-0 pl-[64px] gap-[14.73px] shadow-[0px_0px_5.1px_rgba(0,_0,_0,_0.4)] rounded-[99px] overflow-hidden'>
+                    <div className='font-bold w-[120px] text-main flex items-center'>
+                      전체 리뷰 바로보기
+                    </div>
+                    <div className='w-[75px] h-[75px] relative flex justify-center items-center'>
+                      <Image className='relative' fill alt='' src={allReview} />
+                    </div>
                   </div>
-                  <div className='w-[73px] h-[68px] relative flex justify-center items-center'>
-                    <Image className='relative' fill alt='' src={kakaoReview} />
+                </Link>
+
+                <Link
+                  href={`https://place.map.kakao.com/m/${placeData.kakao_place_id}`}
+                >
+                  <div className='w-[315px] relative h-[45px] flex flex-row items-center justify-start py-0 pl-[64px] gap-[14.73px] shadow-[0px_0px_5.1px_rgba(0,_0,_0,_0.4)] rounded-[99px] overflow-hidden'>
+                    <div className='font-bold w-[132px] text-main flex items-center'>
+                      카카오 리뷰 바로가기
+                    </div>
+                    <div className='w-[73px] h-[68px] relative flex justify-center items-center'>
+                      <Image
+                        className='relative'
+                        fill
+                        alt=''
+                        src={kakaoReview}
+                      />
+                    </div>
                   </div>
-                </div>
-              </Link>
-            </div>
-          </>
-        )}
+                </Link>
+              </div>
+            </>
+          )}
+        </div>
         {contextHolder}
       </div>
     </>
   )
 }
-
-interface SectionProps {
-  ref?: React.RefObject<HTMLDivElement>
-  title: string
-  subtitle?: string
-  button?: React.ReactNode
-  children: React.ReactNode
-}
-
-const Section = React.forwardRef<HTMLDivElement, SectionProps>(
-  ({ title, subtitle, button, children }, ref) => {
-    return (
-      <div className={'w-full flex flex-col gap-[15px] px-[20px]'} ref={ref}>
-        {button ? (
-          <div className={'flex justify-between items-center'}>
-            <span className={'text-main font-semibold'}>{title}</span>
-            {button}
-          </div>
-        ) : (
-          <span className={'text-main font-semibold'}>{title}</span>
-        )}
-        {subtitle && (
-          <span className={'text-sub text-description mt-[-10px]'}>
-            {subtitle}
-          </span>
-        )}
-        {children}
-      </div>
-    )
-  }
-)
-
-Section.displayName = 'Section'
